Update designer styles
