--- src/Designer.css
@@ -1,50 +1,0 @@
-.designer-portrait {
-  width: 100%;
-  border-radius: var(--space-small);
-  margin-bottom: var(--space-medium);
-}
-
-.designer-name {
-  text-align: left;
-  font-weight: 900;
-  font-family: var(--font-family-normal);
-  font-size: var(--font-heading-regular);
-}
-
-.designer-title {
-  text-align: left;
-  font-weight: 100;
-  font-size: var(--font-heading-medium);
-  font-family: var(--font-family-normal);
-}
-
-.designer-bio {
-  text-align: left;
-  font-size: var(--font-regular);
-  margin-bottom: 24px;
-  letter-spacing: 0.1px;
-  line-height: 1.64;
-}
-
-.designer-link {
-  font-size: var(--font-regular);
-}
-
-.designer-contact {
-  text-align: left;
-}
-
-.designer {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  margin-bottom: var(--space-big);
-  padding: var(--space-small);
-}
-
-@media screen and (min-width: 480px) {
-  .designer {
-    min-width: 260px;
-    max-width: 260px;
-  }
-}
