+++ src/Tag.sass
@@ -5,7 +5,7 @@
 @include padding-placeholders($space)
 @include fonts($font-sizes)
 @include colors($color)
-@include weights($weights)
+@include weights($font-weight)
 
 .tag
   @extend %font-tiny
