+++ src/TagList/TagList.sass
@@ -3,7 +3,8 @@
 @include margin-placeholders($space)
 
 .tag-list
-  @extend %margin-vertical-small
+  @extend %margin-top-small
+  @extend %margin-bottom-regular
 
 @media screen and (min-width: 1028px)
   .tag-list
