+++ src/style/_variables.sass
@@ -5,4 +5,3 @@ $font-sizes: (tiny: nth($font-stops, 1), small: nth($font-stops, 2), regular: nt
 $font-family: (sans: Lato, sans-alt: "Open Sans", mono: source-code-pro Menlo monospace)
 $font-weight: (light: 100, normal: 300, bold: 700, black: 900)
 $color: (white: #fff, black: #2a2a2a, gray: #767676)
-$weights: (regular: 400, bold: 700, black: 900)
